+++ src/orderbook/orders/index.ts
@@ -1,5 +1,6 @@
 // Exports
 
+export * as cryptopunks from "@/orderbook/orders/cryptopunks";
 export * as foundation from "@/orderbook/orders/foundation";
 export * as looksRare from "@/orderbook/orders/looks-rare";
 export * as openDao from "@/orderbook/orders/opendao";
@@ -41,7 +42,8 @@ export type OrderKind =
   | "mint"
   | "cryptopunks"
   | "sudoswap"
-  | "universe";
+  | "universe"
+  | "nftx";
 
 // In case we don't have the source of an order readily available, we use
 // a default value where possible (since very often the exchange protocol
@@ -92,6 +94,8 @@ export const getOrderSourceByOrderKind = async (
         return sources.getOrInsert("sudoswap.xyz");
       case "universe":
         return sources.getOrInsert("universe.xyz");
+      case "nftx":
+        return sources.getOrInsert("nftx.io");
       case "mint": {
         if (address && mintsSources.has(address)) {
           return sources.getOrInsert(mintsSources.get(address)!);